+++ sdk/spring/spring-cloud-azure-autoconfigure/src/test/java/com/azure/spring/cloud/autoconfigure/eventhubs/AzureEventHubsTemplateConfigurationTest.java
@@ -4,7 +4,7 @@
 package com.azure.spring.cloud.autoconfigure.eventhubs;
 
 import com.azure.spring.eventhubs.core.EventHubsTemplate;
-import com.azure.spring.eventhubs.core.producer.EventHubProducerFactory;
+import com.azure.spring.eventhubs.core.producer.EventHubsProducerFactory;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.autoconfigure.AutoConfigurations;
 import org.springframework.boot.test.context.FilteredClassLoader;
@@ -16,7 +16,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 class AzureEventHubsTemplateConfigurationTest {
 
     private final ApplicationContextRunner contextRunner = new ApplicationContextRunner()
-        .withConfiguration(AutoConfigurations.of(AzureEventHubMessagingAutoConfiguration.class));
+        .withConfiguration(AutoConfigurations.of(AzureEventHubsMessagingAutoConfiguration.class));
 
     @Test
     void disableEventHubsShouldNotConfigure() {
@@ -25,7 +25,7 @@ class AzureEventHubsTemplateConfigurationTest {
                 "spring.cloud.azure.eventhubs.enabled=false",
                 "spring.cloud.azure.eventhubs.namespace=test-namespace"
             )
-            .run(context -> assertThat(context).doesNotHaveBean(AzureEventHubMessagingAutoConfiguration.EventHubsTemplateConfiguration.class));
+            .run(context -> assertThat(context).doesNotHaveBean(AzureEventHubsMessagingAutoConfiguration.EventHubsTemplateConfiguration.class));
     }
 
     @Test
@@ -35,13 +35,13 @@ class AzureEventHubsTemplateConfigurationTest {
             .withPropertyValues(
                 "spring.cloud.azure.eventhubs.namespace=test-namespace"
             )
-            .run(context -> assertThat(context).doesNotHaveBean(AzureEventHubMessagingAutoConfiguration.EventHubsTemplateConfiguration.class));
+            .run(context -> assertThat(context).doesNotHaveBean(AzureEventHubsMessagingAutoConfiguration.EventHubsTemplateConfiguration.class));
     }
 
     @Test
     void withoutEventHubConnectionShouldNotConfigure() {
         this.contextRunner
-            .run(context -> assertThat(context).doesNotHaveBean(AzureEventHubMessagingAutoConfiguration.EventHubsTemplateConfiguration.class));
+            .run(context -> assertThat(context).doesNotHaveBean(AzureEventHubsMessagingAutoConfiguration.EventHubsTemplateConfiguration.class));
     }
 
     @Test
@@ -50,11 +50,11 @@ class AzureEventHubsTemplateConfigurationTest {
             .withPropertyValues(
                 "spring.cloud.azure.eventhubs.connection-string=" + String.format(CONNECTION_STRING, "test-namespace")
             )
-            .withUserConfiguration(AzureEventHubPropertiesTestConfiguration.class)
+            .withUserConfiguration(AzureEventHubsPropertiesTestConfiguration.class)
             .run(context -> {
                 assertThat(context).hasSingleBean(EventHubsTemplate.class);
-                assertThat(context).hasSingleBean(EventHubProducerFactory.class);
-                assertThat(context).hasSingleBean(AzureEventHubMessagingAutoConfiguration.EventHubsTemplateConfiguration.class);
+                assertThat(context).hasSingleBean(EventHubsProducerFactory.class);
+                assertThat(context).hasSingleBean(AzureEventHubsMessagingAutoConfiguration.EventHubsTemplateConfiguration.class);
             });
     }